pyproject.toml
[tool.poetry]
name = "django-account"
version = "0.1.0"
description = "django account"
authors = ["Hasbi Asshidiq <hasbiasshidiq@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
pip = "^24.0"
install = "^1.3.5"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
